+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 seleniumtid changelog
 =====================
 
+v0.8
+----
+
+*Release date: 2015-01-??*
+
+  * Add visual testing to SeleniumTestCase  
+    New config property 'visualtests_enabled' in [Server] section to compare screenshots with the baseline images  
+    New config property 'visualtests_save' in [Server] section to save the screenshots
+
 v0.7
 ----